Refactored code and configuration
for stream and event processing
