+++ src/main/java/io/karthik/cepengine/processors/KafkaStreamsIOInterface.java
@@ -1,12 +1,16 @@
-package io.karthik.cepengine.streamers;
+package io.karthik.cepengine.processors;
 
 import org.apache.kafka.streams.kstream.KStream;
 import org.springframework.cloud.stream.annotation.Input;
 
 public interface KafkaStreamsIOInterface {
 
-  String INPUT = "shipment";
+  String INPUT = "kstream-input-topic";
+  String OUTPUT = "kstream-output-topic";
 
   @Input(INPUT)
   KStream<?, ?> input();
+
+//  @Output(OUTPUT)
+//  KStream<?, ?> output();
 }
