+++ src/main/java/io/karthik/cepengine/processors/stream/ShipmentStreamProcessor.java
@@ -1,5 +1,6 @@
-package io.karthik.cepengine.streamers;
+package io.karthik.cepengine.processors.stream;
 
+import io.karthik.cepengine.processors.KafkaStreamsIOInterface;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.kafka.streams.kstream.KStream;
 import org.slf4j.Logger;
@@ -9,12 +10,12 @@ import org.springframework.cloud.stream.annotation.EnableBinding;
 import org.springframework.cloud.stream.annotation.StreamListener;
 import org.springframework.stereotype.Component;
 
-@ConditionalOnProperty(name = "cep-properties.processors.KStreamShipmentStreamer.enabled", havingValue = "true")
-@EnableBinding(KafkaStreamsIOInterface.class)
 @Component
-public class KStreamShipmentStreamer {
+@EnableBinding(KafkaStreamsIOInterface.class)
+@ConditionalOnProperty(name = "cep-properties.processors.kstream-shipment-streamer.enabled", havingValue = "true")
+public class ShipmentStreamProcessor {
 
-  private final Logger log = LoggerFactory.getLogger(KStreamShipmentStreamer.class);
+  private final Logger log = LoggerFactory.getLogger(ShipmentStreamProcessor.class);
 
   @StreamListener(KafkaStreamsIOInterface.INPUT)
   public void processStream(KStream<GenericRecord, GenericRecord> inputStream) {
