+++ src/main/resources/application.yaml
@@ -1,72 +1,87 @@
 # Server Config
 server:
   port: 9009
-
 #Spring config
 spring:
   cloud:
     stream:
-      # Kafka Binder configuration
-      kafka:
-        binder:
-          brokers: localhost
-          defaultBrokerPort: 9093
-          producer-properties:
-            client.id: Cep-Engine
-            schema.registry.url: http://localhost:8081
-            requiredAcks: 1
-            value.serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
-            key.serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
-          consumer-properties:
-            schema.registry.url: http://localhost:8081
-            value.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
-            key.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
-        # Streams binder configuration
-        streams:
-          binder:
-            brokers: localhost:9093
-            configuration:
-              commit.interval.mms: 1000
-              schema.registry.url: http://localhost:8081
-          bindings:
-            shipment: # name of the topic
-              consumer:
-                applicationId: cep-engine-kstreams
-                valueSerde: io.confluent.kafka.streams.serdes.avro.GenericAvroSerde
-                keySerde: io.confluent.kafka.streams.serdes.avro.GenericAvroSerde
-              group: shipment-consumer-streamer
-
-      # TODO: Enable when needed
-      ## Kafka event processin config
-      ### Defining the binders involved
+      ## Defining the binders involved
       binders:
+        # Kafka binder configuration
         kafka:
           type: kafka
+          environment:
+            spring:
+              cloud:
+                stream:
+                  kafka:
+                    binder:
+                      brokers: localhost
+                      defaultBrokerPort: 9093
+                      producer-properties:
+                        client.id: Cep-Engine
+                        schema.registry.url: http://localhost:8081
+                        requiredAcks: 1
+                        value.serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
+                        key.serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
+                      consumer-properties:
+                        schema.registry.url: http://localhost:8081
+                        value.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+                        key.deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+        # KStreams binder configuration
+        kstream:
+          type: kstream
+          environment:
+            spring:
+              cloud:
+                stream:
+                  kafka:
+                    streams:
+                      binder:
+                        brokers: localhost:9093
+                        configuration:
+                          commit.interval.mms: 1000
+                          schema.registry.url: http://localhost:8081
+                          default.key.serde: io.confluent.kafka.streams.serdes.avro.GenericAvroSerde
+                          default.value.serde: io.confluent.kafka.streams.serdes.avro.GenericAvroSerde
+                          group.id: shipment-stream-consumer
       # Binding for input and output channels
       bindings:
-        InputTopic:
+        # Event processing input topic
+        input-topic:
           destination: shipment
-          group: shipment-consumer
+          group: shipment-event-consumer
           content-type: application/*+avro
           binder: kafka
-        OutputTopic:
+        # Event processing output topic
+        output-topic:
           destination: shipment-anomaly
           group: shipment-filtered-producer
           content-type: application/*+avro
           binder: kafka
           producer:
             use-native-encoding: true
-
+        # Stream processing input topic
+        kstream-input-topic:
+          destination: shipment
+          group: shipment-stream-consumer
+          binder: kstream
+          applicationId: shipment-stream-consumer
+        # Stream processing output topic
+        kstream-output-topic:
+          destination: shipment-stream-anomaly-only
+          group: shipment-stream-consumer
+          binder: kstream
+          applicationId: shipment-stream-producer
 # Logging Config
 logging:
   level:
     root: INFO
-
 # Custom properties
 cep-properties:
-  schema-location: ${SCHEMA_LOCATION}
   processors:
-    ShipmentStreamer:
+    shipment-event-transformer:
       enabled: true
-    KStreamShipmentStreamer:
+      output-topic-schema-location: ${SHIPMENT_EVENT_OUTPUT_SCHEMA_LOCATION}
+    kstream-shipment-streamer:
       enabled: true
